+++ src/features/Home/index.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import TimelineLayout from '@/components/Timeline/TimelineLayout'
 import Timeline from '@/components/Timeline/Timeline'
+import { SubscriptionProvider } from '@/context/SubscriptionContext'
 
 interface HomePageProps {
   focusBottomTab: () => void
@@ -15,12 +16,14 @@ const HomePage: React.FC<HomePageProps> = ({
 }) => {
   return (
     <TimelineLayout>
-      <Timeline
-        onScrollUp={focusBottomTab}
-        onScrollDown={unfocusBottomTab}
-        onToggleFollow={toggleFollow}
-        showTabs={true}
-      />
+      <SubscriptionProvider>
+        <Timeline
+          onScrollUp={focusBottomTab}
+          onScrollDown={unfocusBottomTab}
+          onToggleFollow={toggleFollow}
+          showTabs={true}
+        />
+      </SubscriptionProvider>
     </TimelineLayout>
   )
 }
